+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Improved live generator responses
+
 ## v0.2.5 (2023-11-08)
 * Fixed end-of-stack Exception handling